requirements.txt
wheel
urllib3==1.26.8
jsonpath-rw-ext==1.2.2
python-stdnum==1.9
rfc3987==1.3.8
requests==2.27.1
python-dateutil==2.8.0
Flask==2.0.2
Flask-Cors==3.0.10
PyPika==0.48.8
lxml==4.7.1
psutil==5.9.0
gunicorn==20.1.0
orjson==3.6.7
mysql-connector-python==8.0.28
pytest==7.0.1
mock>=1.0.1
python-Levenshtein==0.12.1
aenum==3.1.11
pathvalidate==2.5.2
flexmock==0.11.3
pycld3==0.22
